feat: Add application service for processing late jobs
