+++ api/src/application/services.rs
@@ -2,5 +2,6 @@ pub mod create_monitor;
 pub mod delete_monitor;
 pub mod fetch_job;
 pub mod finish_job;
+pub mod process_late_jobs;
 pub mod start_job;
 pub mod update_monitor;
